// AngularMain/src/app/modules/models/FunctionMst.ts
    export interface FunctionMst {
        tenentID: number;
        menU_ID: number;
        masteR_ID: number;
        modulE_ID: number;
        menU_TYPE: string;
        MENU_NAMEEnglish: string;
        menU_NAMEArabic: string;
        menU_NAME3: string;
        link: string;
        urloption: string;
        urlrewrite: string;
        menU_LOCATION: string;
        menU_ORDER: string;
        doC_PARENT: string;
        cruP_ID: number;
        addflage: string;
        editflage: string;
        delflage: string;
        printflage: string;
        amiglobale: string;
        mypersonal: string;
        smalltext: string;
        activetilldate: string;
        iconpath: string;
        commanline: string;
        activE_FLAG: string;
        metatitle: string;
        metakeyword: string;
        metadescription: string;
        headervisibledata: string;
        headerinvisibledata: string;
        footervisibledata: string;
        footerinvisibledata: string;
        refid: number;
        mybusid: number;
        lableflag: number;
        sP1: number;
        sP2: number;
        sP3: number;
        sP4: number;
        sP5: number;
        sP1Name: string;
        sP2Name: string;
        sP3Name: string;
        sP4Name: string;
        sP5Name: string;
        activemenu: number;
        menudate: number;
    }